+++ .formatter.exs
@@ -4,5 +4,9 @@
     :ecto,
     :ecto_sql
   ],
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
+  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"],
+  export: [
+    locals_without_parens: [assert_enqueued: 1, refute_enqueued: 1]
+  ],
+  locals_without_parens: [assert_enqueued: 1, refute_enqueued: 1]
 ]